refactor out BoardDocs mixin
